+++ lib/src/LogRecordFormatter.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:basic_utils/basic_utils.dart';
 import 'package:intl/intl.dart';
 import 'package:log_4_dart_2/log_4_dart_2.dart';
 
@@ -7,15 +8,25 @@ import '../log_4_dart_2.dart';
 
 class LogRecordFormatter {
   ///
-  /// Converts the given [logRecord] to the given [format]
+  /// Converts the given [logRecord] to the given [format].
   ///
-  static String format(LogRecord logRecord, String format) {
+  /// The [dateFormat] defines the format for the [LogRecord.time]
+  ///
+  static String format(LogRecord logRecord, String format,
+      {String dateFormat = 'yyyy-MM-dd HH:mm:ss'}) {
     if (format.contains('\%d')) {
-      var date = DateFormat('yyyy-MM-dd HH:mm:ss').format(logRecord.time);
+      var date = DateFormat(dateFormat).format(logRecord.time);
       format = format.replaceAll('\%d', date);
     }
     if (format.contains('\%t')) {
       format = format.replaceAll('\%t', logRecord.loggerName);
+    }
+    if (format.contains('\%i')) {
+      if (StringUtils.isNullOrEmpty(logRecord.identifier)) {
+        format = format.replaceAll('\%i', '');
+      } else {
+        format = format.replaceAll('\%i', logRecord.identifier);
+      }
     }
     if (format.contains('\%l')) {
       format = format.replaceAll('\%l', logRecord.level.name);
@@ -29,9 +40,12 @@ class LogRecordFormatter {
   ///
   /// Converts the given [logRecord] to a json string for the [HttpAppender].
   ///
-  static String formatJson(LogRecord logRecord) {
+  /// The [dateFormat] defines the format for the [LogRecord.time]
+  ///
+  static String formatJson(LogRecord logRecord,
+      {String dateFormat = 'yyyy-MM-dd HH:mm:ss'}) {
     var map = {
-      'time': DateFormat('yyyy-MM-dd HH:mm:ss').format(logRecord.time),
+      'time': DateFormat(dateFormat).format(logRecord.time),
       'message': logRecord.message,
       'level': logRecord.level.toString(),
       'tag': logRecord.loggerName,