+++ lib/src/appender/Appender.dart
@@ -5,23 +5,29 @@ import 'package:log_4_dart_2/src/appender/AppenderType.dart';
 /// The basic appender.
 ///
 abstract class Appender {
-  /// The date the appender was created
+  /// The date the appender was created.
   DateTime created;
 
-  /// The loglevel for the appender
+  /// The loglevel for the appender.
   Level level;
 
-  /// The logformat for the appender
+  /// The logformat for the appender.
   String format;
 
-  /// The type of the appender
+  /// The type of the appender.
   AppenderType type;
 
-  /// The default logging format
+  /// The default logging format.
   static String defaultFormat = '%d %t %l %m';
 
+  /// The default date format.
+  static String defaultDateFormat = 'yyyy-MM-dd HH:mm:ss';
+
+  /// The dateformat used for the appender.
+  String dateFormat;
+
   ///
-  /// Appending the given [logRecord]
+  /// Appending the given [logRecord].
   ///
   void append(LogRecord logRecord);
 
@@ -29,4 +35,14 @@ abstract class Appender {
   /// Setup the appender. This needs to be called for every appender to configure the appender with the necessary data.
   ///
   void init(Map<String, dynamic> config, bool test, DateTime date);
+
+  ///
+  /// Retuns a new instance of the appender.
+  ///
+  Appender getInstance();
+
+  ///
+  /// Returns the type of the appender.
+  ///
+  String getType();
 }