+++ test/log_record_formatter_test.dart
@@ -5,15 +5,22 @@ import 'package:test/test.dart';
 
 void main() {
   test('Test format()', () {
-    var record = LogRecord(Level.INFO, 'Lorem Ipsum', 'TestClass');
+    var record =
+        LogRecord(Level.INFO, 'Lorem Ipsum', 'TestClass', identifier: 'uuid');
+
     var now = DateTime.now();
     record.time = now;
-    var formatted = LogRecordFormatter.format(record, '%d %t %l %m');
+    var formatted = LogRecordFormatter.format(record, '%d %i %t %l %m');
     expect(formatted,
-        '${DateFormat('yyyy-MM-dd HH:mm:ss').format(now)} TestClass INFO Lorem Ipsum');
+        '${DateFormat('yyyy-MM-dd HH:mm:ss').format(now)} uuid TestClass INFO Lorem Ipsum');
 
-    formatted = LogRecordFormatter.format(record, '%d %t %l: %m');
+    formatted = LogRecordFormatter.format(record, '%d %i %t %l: %m');
     expect(formatted,
-        '${DateFormat('yyyy-MM-dd HH:mm:ss').format(now)} TestClass INFO: Lorem Ipsum');
+        '${DateFormat('yyyy-MM-dd HH:mm:ss').format(now)} uuid TestClass INFO: Lorem Ipsum');
+
+    formatted = LogRecordFormatter.format(record, '%d %i %t %l: %m',
+        dateFormat: 'yyyy-MM-dd');
+    expect(formatted,
+        '${DateFormat('yyyy-MM-dd').format(now)} uuid TestClass INFO: Lorem Ipsum');
   });
 }